// cloudreve-official-frontend/src/api/dashboard.ts
export interface QQConnectConfig {
  app_id?: string;
  app_secret?: string;
  direct_sign_in?: boolean;
}

export interface LogtoConfig {
  endpoint?: string;
  app_id?: string;
  app_secret?: string;
  direct_sign_in?: boolean;
  display_name?: string;
  icon?: string;
  direct_sso?: string;
}

export interface OIDCConfig {
  wellknown_endpoint?: string;
  client_id?: string;
  client_secret?: string;
  scope?: string;
  direct_sign_in?: boolean;
  display_name?: string;
  sub_field_mapping?: string;
  email_field_mapping?: string;
  avatar_field_mapping?: string;
  name_field_mapping?: string;
  icon?: string;
}